refactor: extract write/COW path into dev/write.rs
Move the write path (ensure_l2_offset, the COW helpers, write-mapping
construction and write_at) into a new dev/write.rs submodule as an impl
extension of Qcow2Dev. No visibility changes were needed: it calls the
already-pub(crate) read/alloc helpers and the parent's private cache
kernel, and exposes only the existing pub write_at. No logic changes.
